requirements.txt
requests>=2.22,<2.23
python-dotenv==0.13.0